+++ src/ReverseProxy/Service/RuntimeModel/ClusterConfig.cs
@@ -62,104 +62,5 @@ namespace Microsoft.ReverseProxy.RuntimeModel
         {
             return !Cluster.Equals(_cluster, newClusterConfig._cluster);
         }
-
-        /// <summary>
-        /// All health check options for a cluster.
-        /// </summary>
-        /// <remarks>
-        /// Struct used only to keep things organized as we add more configuration options inside of `ClusterConfig`.
-        /// Each "feature" can have its own struct.
-        /// </remarks>
-        public readonly struct ClusterHealthCheckOptions
-        {
-            public ClusterHealthCheckOptions(ClusterPassiveHealthCheckOptions passive, ClusterActiveHealthCheckOptions active)
-            {
-                Passive = passive;
-                Active = active;
-            }
-
-            /// <summary>
-            /// Whether at least one type of health check is enabled.
-            /// </summary>
-            public bool Enabled => Passive.Enabled || Active.Enabled;
-
-            /// <summary>
-            /// Passive health check options.
-            /// </summary>
-            public ClusterPassiveHealthCheckOptions Passive { get; }
-
-            /// <summary>
-            /// Active health check options.
-            /// </summary>
-            public ClusterActiveHealthCheckOptions Active { get; }
-        }
-
-        /// <summary>
-        /// Passive health check options for a cluster.
-        /// </summary>
-        public readonly struct ClusterPassiveHealthCheckOptions
-        {
-            public ClusterPassiveHealthCheckOptions(bool enabled, string policy, TimeSpan? reactivationPeriod)
-            {
-                Enabled = enabled;
-                Policy = policy;
-                ReactivationPeriod = reactivationPeriod;
-            }
-
-            /// <summary>
-            /// Whether active health checks are enabled.
-            /// </summary>
-            public bool Enabled { get; }
-
-            /// <summary>
-            /// Passive health check policy.
-            /// </summary>
-            public string Policy { get; }
-
-            /// <summary>
-            /// Destination reactivation period after which an unhealthy destination is considered healthy again.
-            /// </summary>
-            public TimeSpan? ReactivationPeriod { get; }
-        }
-
-        /// <summary>
-        /// Active health check options for a cluster.
-        /// </summary>
-        public readonly struct ClusterActiveHealthCheckOptions
-        {
-            public ClusterActiveHealthCheckOptions(bool enabled, TimeSpan? interval, TimeSpan? timeout, string policy, string path)
-            {
-                Enabled = enabled;
-                Interval = interval;
-                Timeout = timeout;
-                Policy = policy;
-                Path = path;
-            }
-
-            /// <summary>
-            /// Whether active health checks are enabled.
-            /// </summary>
-            public bool Enabled { get; }
-
-            /// <summary>
-            /// Health probe interval.
-            /// </summary>
-            public TimeSpan? Interval { get; }
-
-            /// <summary>
-            /// Health probe timeout, after which a destination is considered unhealthy.
-            /// </summary>
-            public TimeSpan? Timeout { get; }
-
-            /// <summary>
-            /// Active health check policy.
-            /// </summary>
-            public string Policy { get; }
-
-            /// <summary>
-            /// HTTP health check endpoint path.
-            /// </summary>
-            public string Path { get; }
-        }
     }
 }
