+++ test/ReverseProxy.Tests/Middleware/LoadBalancerMiddlewareTests.cs
@@ -96,7 +96,7 @@ namespace Microsoft.ReverseProxy.Middleware.Tests
             var httpClient = new HttpMessageInvoker(new Mock<HttpMessageHandler>().Object);
             var cluster1 = new ClusterInfo(
                 clusterId: "cluster1",
-                destinationManager: new DestinationManager(null));
+                destinationManager: new DestinationManager());
             cluster1.ConfigSignal.Value = new ClusterConfig(default, default, new ClusterLoadBalancingOptions(LoadBalancingMode.RoundRobin), default, httpClient, default, new Dictionary<string, string>());
             var destination1 = cluster1.DestinationManager.GetOrCreateItem(
                 "destination1",
